requirements.txt
google-adk[a2a] 
google-genai 
python-dotenv
uvicorn
a2a-sdk
requests